modified function comments for JSDOC and removed mapStringToNumbers
