+++ jsdocTags.js
@@ -1,0 +1,4 @@
+exports.defineTags = function (dictionary) {
+    dictionary.defineTag('server');
+    dictionary.defineTag('client');
+};
