+++ server/lib/composeUpdateQuestionStatement.js
@@ -1,5 +1,5 @@
 /**
- * Represents a function that creates the prepared object required to update questions
+ * Represents a function that composes a statement to update (n) number of questiosn in the database
  * @param {array} rows - array of objects that need to be saved to the database
  * @param {function} callback - a callback function
  */
@@ -30,10 +30,6 @@ function composeUpdateQuestionStatement (rows, callback) {
         valueClause.push('$' + params.length);
         chunks.push('(' + valueClause.join(', ') + ')');
     }
-
-    // INSERT INTO questions (question_id, quiz_id, question, a, b, c, d, correct_answer) VALUES (1, 1, 'What is the capital of Japan?', 'London', 'Cardiff', 'Tokyo', 'Doncaster', 'C') ON CONFLICT (question_id) DO UPDATE SET question = EXCLUDED.question, a = EXCLUDED.a, b = EXCLUDED.b, c = EXCLUDED.c, d = EXCLUDED.d, correct_answer = EXCLUDED.correct_answer;
-
-
 
     callback(null, {
         text: 'INSERT INTO questions (question_id, quiz_id, question, a, b, c, d, correct_answer) VALUES ' + chunks.join(', ') + 'ON CONFLICT (question_id) DO UPDATE SET question = EXCLUDED.question, a = EXCLUDED.a, b = EXCLUDED.b, c = EXCLUDED.c, d = EXCLUDED.d, correct_answer = EXCLUDED.correct_answer;',
