+++ server/lib/getQuizMembers.js
@@ -1,7 +1,7 @@
 var query = require('./query');
 
 /**
- * Represents a function that fetches a list of users belonging to a quiz
+ * Represents a function that returns a list of students belonging to a quiz
  * @param {object} client - postgres database client
  * @param {number} quiz_id - quiz id to which all students will belong to
  * @param {function} callback - a callback function
