+++ server/lib/getUserByEmail.js
@@ -1,12 +1,12 @@
 var query = require('./query');
 
 /**
- * Represents a function that returns user details
+ * Represents a function that returns user details by email
  * @param {object} client - postgres database client
  * @param {string} email - email for the given user
  * @param {function} callback - a callback function
  */
-function getUser (client, email, callback) {
+function getUserByEmail (client, email, callback) {
 
     var userQuery = 'SELECT * FROM users WHERE email = $1;';
     var userValue = [email];
@@ -20,4 +20,4 @@ function getUser (client, email, callback) {
     });
 }
 
-module.exports = getUser;
+module.exports = getUserByEmail;
