+++ server/lib/validateModuleID.js
@@ -7,6 +7,7 @@ var query = require('./query');
  * @param {string} module_id - module_id
  * @param {function} callback - a callback function
  */
+ 
 function validateModuleID (client, module_id, callback) {
 
     var moduleQuery = 'SELECT EXISTS(SELECT 1 FROM modules WHERE module_id = $1);';
