+++ src/js/lib/getUserID.js
@@ -3,7 +3,6 @@
  * Returns the user ID (string) if the cookie exists, or undefined.
  */
 
-
 export default function getUserID () {
     try {
         const result = document.cookie.match(/cul_id=\d+/)[0].match(/\d+/);
