+++ src/js/lib/isUserLecturer.js
@@ -3,7 +3,6 @@
  * Returns the is_lecturer (boolean) if the cookie exists, or undefined.
  */
 
-
 export default function isUserLecturer () {
     try {
         const result = document.cookie.match(/cul_is_lecturer=(true|false)/)[0].match(/(true|false)/);
