+++ src/js/lib/validCookieExists.js
@@ -1,9 +1,7 @@
 /**
- * Represents a function that Checks for the existence of a valid cookie
+ * Represents a function that checks for the existence of a valid cookie
  */
 export default function validCookieExists () {
 
     return document.cookie.indexOf("cul_id") !== -1;
 }
-
-// is there any point to this function if getUserID exists?
