+++ test/unit/server/getUser.test.js
@@ -1,14 +1,14 @@
 import test from 'tape';
-import getUser from '../../../server/lib/getUser';
+import getUserByEmail from '../../../server/lib/getUserByEmail';
 import { testClient } from '../../utils/init';
 import { users as expectedRows } from '../../utils/data-fixtures';
 
-test('`getUser` works', (t) => {
+test('`getUserByEmail` works', (t) => {
 
     t.plan(1);
     const emailAddress = 'student@city.ac.uk';
 
-    getUser(testClient, emailAddress, (error, response) => {
+    getUserByEmail(testClient, emailAddress, (error, response) => {
 
         if (error) {
             console.error(error);
